+++ src/EluvioLive.js
@@ -354,199 +354,6 @@ class EluvioLive {
   }
 
   /**
-   * Return tenant admins group and content admins group corresponding to this tenant.
-   * @param {string} tenantId - The ID of the tenant (iten***)
-   */
-  async TenantShowNew({ tenantId }) {
-    let tenantInfo = {};
-
-    const tenantAddr = Utils.HashToAddress(tenantId);
-    const abi = fs.readFileSync(
-      path.resolve(__dirname, "../contracts/v3/BaseTenantSpace.abi")
-    );
-
-    let errors = [];
-
-    let tenantAdminAddr = await this.client.CallContractMethod({
-      contractAddress: tenantAddr,
-      abi: JSON.parse(abi),
-      methodName: "groupsMapping",
-      methodArgs: ["tenant_admin", 0],
-      formatArguments: true,
-    });
-    tenantInfo["tenant_admin_address"] = tenantAdminAddr;
-    
-    //Content admins group might not exist for the tenant with this tenantId due to legacy reasons.
-    //Running ./elv-live tenant-fix to update this tenant.
-    let contentAdminAddr;
-    try {
-      contentAdminAddr = await this.client.CallContractMethod({
-        contractAddress: tenantAddr,
-        abi: JSON.parse(abi),
-        methodName: "groupsMapping",
-        methodArgs: ["content_admin", 0], 
-        formatArguments: true,
-      });
-    } catch (e) {
-      contentAdminAddr = null;
-      errors.push("missing content admins");
-    }
-    tenantInfo["content_admin_address"] = contentAdminAddr;
-    console.log(await this.client.authClient.AccessType("iten"+Utils.AddressToHash(contentAdminAddr)));
-
-    //Check if the groups have _ELV_TENANT_ID set correctly
-    for (const group in tenantInfo) {
-      let groupAddress = tenantInfo[group];
-      
-      if (groupAddress != null) {
-        let tenantContractAddress = await this.client.CallContractMethod({
-          contractAddress: groupAddress,
-          methodName: "getMeta",
-          methodArgs: ["_ELV_TENANT_ID"], 
-        });
-
-        let args = group.split("_");
-        if (tenantContractAddress == "0x") {
-          errors.push(`${args[0]} ${args[1]} group is not associated with any tenant`);
-        } else {
-          let tenantContractId = Ethers.utils.toUtf8String(tenantContractAddress);
-
-          if (tenantContractId != tenantId) {
-            errors.push(`${args[0]} ${args[1]} group doesn't belong to this tenant`);
-          }
-        }
-      }
-    }
-
-    if (errors.length != 0) {
-      tenantInfo["errors"] = errors;
-    }
-
-    return tenantInfo;
-  }
-
-  /**
-   * Create a new content admin group corresponding to this tenant.
-   * @param {string} tenantId - The ID of the tenant (iten***)
-   * @returns {string} Content Admin Group's address
-   */
-  async TenantSetContentAdmins({ tenantId }) {
-    //Check that the user is the owner of the tenant
-    const tenantOwner = await this.client.authClient.Owner({id: tenantId});
-    if (tenantOwner.toLowerCase() != this.client.signer.address.toLowerCase()) {
-      throw Error("Content Admin must be created by the owner of tenant " + tenantId);
-    }
-
-    //The tenant must not already have a content admin group - can only have 1 content admin group for each tenant.
-    const abi = fs.readFileSync(
-      path.resolve(__dirname, "../contracts/v3/BaseTenantSpace.abi")
-    );
-
-    const tenantAddr = Utils.HashToAddress(tenantId);
-
-    let contentAdmin;
-    try {
-      contentAdmin = await this.client.CallContractMethod({
-        contractAddress: tenantAddr,
-        abi: JSON.parse(abi),
-        methodName: "groupsMapping",
-        methodArgs: ["content_admin", 0],
-        formatArguments: true,
-      });
-    } catch (e) {
-      //call cannot override gasLimit error will be thrown if content admin group doesn't exist for this tenant.
-      contentAdmin = null;
-    }
-
-    if (contentAdmin) {
-      let logMsg = `Tenant ${tenantId} already has a content admin group: ${contentAdmin}, aborting...`;
-      return logMsg;
-    }
-
-    //Create and add the new content admin group to the user's account.
-    let elvAccount = new ElvAccount({configUrl:this.configUrl, debugLogging: this.debug});
-    elvAccount.InitWithClient({elvClient:this.client});
-
-    let accountName = await this.client.userProfileClient.UserMetadata({
-      metadataSubtree: "public/name"
-    });
-  
-    let contentAdminGroup = await elvAccount.CreateAccessGroup({
-      name: `${accountName} Content Admins`,
-    });
-
-    await elvAccount.AddToAccessGroup({
-      groupAddress: contentAdminGroup.address,
-      accountAddress: this.client.signer.address.toLowerCase(),
-      isManager: true,
-    });
-
-    //Associate the group with this tenant - set the content admin group's _ELV_TENANT_ID to this tenant's tenant id.
-    await this.GroupSetTenantContractId({tenantId: tenantId, groupAddress: contentAdminGroup.address});
-
-    //Associate the tenant with this group - set tenant's content admin group on the tenant's contract.
-    let res = await this.client.CallContractMethodAndWait({
-      contractAddress: tenantAddr,
-      abi: JSON.parse(abi),
-      methodName: "addGroup",
-      methodArgs: ["content_admin", contentAdminGroup.address],
-      formatArguments: true,
-    });
-
-    return res;
-  }
-
-  /**
-   * Remove a content admin from this tenant.
-   * @param {string} tenantId - The ID of the tenant (iten***)
-   * @param {string} contentAdminsAddress - Address of content admin we want to remove.
-   */
-  async TenantRemoveContentAdmin({ tenantId, contentAdminsAddress }) {
-    const abi = fs.readFileSync(
-      path.resolve(__dirname, "../contracts/v3/BaseTenantSpace.abi")
-    );
-
-    const tenantAddr = Utils.HashToAddress(tenantId);
-
-    let res = await this.client.CallContractMethodAndWait({
-      contractAddress: tenantAddr,
-      abi: JSON.parse(abi),
-      methodName: "removeGroup",
-      methodArgs: ["content_admin", contentAdminsAddress],
-      formatArguments: true,
-    });
-
-    return res;
-  }
-
-  /**
-   * Set _ELV_TENANT_ID of group with groupAddress to tenantId.
-   * @param {string} tenantId - The ID of the tenant (iten***)
-   * @param {string} groupAddress - Address of the group we want to remove.
-   */
-  async GroupSetTenantContractId({ tenantId, groupAddress }) {
-    let id = await this.client.CallContractMethod({
-      contractAddress: groupAddress,
-      methodName: "getMeta",
-      methodArgs: ["_ELV_TENANT_ID"], 
-    });
-    if (id != "0x") {
-      throw Error(`Group ${groupAddress} already has a _ELV_TENANT_ID metadata set to ${id}, aborting...`);
-    }
-
-    let newId = await this.client.CallContractMethod({
-      contractAddress: groupAddress,
-      methodName: "putMeta",
-      methodArgs: [
-        "_ELV_TENANT_ID",
-        tenantId
-      ],
-    });
-
-    return newId;
-  }
-
-  /**
    * Get a list of the NFTs of this tenant owned by 'ownerAddr'
    *
    * @namedParams
