+++ utilities/EluvioLiveCli.js
@@ -292,102 +292,6 @@ const CmdTenantShow = async ({ argv }) => {
   }
 };
 
-const CmdTenantShowNew = async({ argv }) => {
-  console.log("Tenant - show", argv.tenant);
-  try {
-    await Init({ debugLogging: argv.verbose, asUrl: argv.as_url });
-
-    let res = await elvlv.TenantShowNew({
-      tenantId: argv.tenant
-    });
-
-    console.log(yaml.dump(res));
-    if (res.errors) { 
-      console.log(`ERROR: tenant_show detected ${res.errors.length} error(s), run ./elv-live tenant_fix ${argv.tenant} to resolve them.`);
-    }
-  } catch (e) {
-    console.error("ERROR:", e);
-  }
-};
-
-const CmdTenantFix = async({ argv }) => {
-  console.log("Tenant - fix", argv.tenant);
-  try {
-    await Init({ debugLogging: argv.verbose, asUrl: argv.as_url });
-
-    let res = await elvlv.TenantShowNew({
-      tenantId: argv.tenant
-    });
-    let errors = res.errors;
-    let unresolved = [];
-
-    if (!errors) {
-      console.log(`No error found for tenant with tenant id: ${argv.tenant}!`);
-      return;
-    }
-
-    for (let i = 0; i < errors.length; i++) {
-      let error = errors[i];
-      switch(error) {
-        case 'missing content admins':
-          let content_admin_address = await elvlv.TenantSetContentAdmins({tenantId: argv.tenant});
-          console.log(`Created new content admin group ${content_admin_address} for tenant with tenantId ${argv.tenant}`);
-          break;
-
-        case 'tenant admin group is not associated with any tenant': 
-          await elvlv.GroupSetTenantContractId({tenantId: argv.tenant, groupAddress: res.tenant_admin_address});
-          break;
-
-        case 'content admin group is not associated with any tenant':
-          await elvlv.GroupSetTenantContractId({tenantId: argv.tenant, groupAddress: res.content_admin_address});
-          break;
-
-        default:
-          console.log(`No fix actions available for error: ${error}.`)
-          unresolved.push(error);
-      }
-    }
-
-    console.log(`${errors.length - unresolved.length} errors fixed, ${unresolved.length} errors remaining.`)
-  } catch (e) {
-    console.error("ERROR:", e);
-  }
-}
-
-const CmdTenantSetContentAdmins = async ({ argv }) => {
-  console.log(`Setting a new content admin group for Tenant ${argv.tenant}`);
-
-  try {
-    await Init({ debugLogging: argv.verbose, asUrl: argv.as_url });
-
-    let res = await elvlv.TenantSetContentAdmins({
-      tenantId: argv.tenant,
-    });
-
-    console.log(yaml.dump(res));
-  } catch (e) {
-    console.error("ERROR:", e);
-  }
-};
-
-const CmdTenantRemoveContentAdmin = async ({ argv }) => {
-  console.log(`Removing a Content Admin from Tenant ${argv.tenant}`);
-  console.log(`Removed Content Admin's Address: ${argv.content_admin_address}`);
-  
-  try {
-    await Init({ debugLogging: argv.verbose, asUrl: argv.as_url });
-
-    let res = await elvlv.TenantRemoveContentAdmin({
-      tenantId: argv.tenant,
-      contentAdminsAddress: argv.content_admin_address
-    });
-
-    console.log(yaml.dump(res));
-  } catch (e) {
-    console.error("ERROR:", e);
-  }
-};
-
 const CmdTenantSetTokenURI = async ({ argv }) => {
   console.log("request_type ", argv.request_type);
   console.log("tenantId ", argv.tenant);
@@ -2162,70 +2066,6 @@ yargs(hideBin(process.argv))
     },
     (argv) => {
       CmdTenantShow({ argv });
-    }
-  )
-
-  .command(
-    "tenant_show_new <tenant> [options]",
-    "Show info on this tenant",
-    (yargs) => {
-      yargs
-        .positional("tenant", {
-          describe: "Tenant ID",
-          type: "string",
-        });
-    },
-    (argv) => {
-      CmdTenantShowNew({ argv });
-    }
-  )
-
-  .command(
-    "tenant_fix <tenant> [options]",
-    "Fix tenant",
-    (yargs) => {
-      yargs
-        .positional("tenant", {
-          describe: "Tenant ID",
-          type: "string",
-        });
-    },
-    (argv) => {
-      CmdTenantFix({ argv });
-    }
-  )
-
-  .command(
-    "tenant_set_content_admins <tenant>",
-    "Set new content admin",
-    (yargs) => {
-      yargs
-        .positional("tenant", {
-          describe: "Tenant ID",
-          type: "string",
-        });
-    },
-    (argv) => {
-      CmdTenantSetContentAdmins({ argv });
-    }
-  )
-
-  .command(
-    "tenant_remove_content_admin <tenant> <content_admin_address>",
-    "Remove a content admin",
-    (yargs) => {
-      yargs
-        .positional("tenant", {
-          describe: "Tenant ID",
-          type: "string",
-        })
-        .positional("content_admin_address", {
-          describe: "Content Admin's address",
-          type: "string",
-        });
-    },
-    (argv) => {
-      CmdTenantRemoveContentAdmin({ argv });
     }
   )
